requirements.txt
cryptography>=41.0.0
pyOpenSSL>=23.2.0
requests>=2.31.0
sqlalchemy>=2.0.0
alembic>=1.12.0
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0
celery>=5.3.0
redis>=4.6.0
python-crontab>=3.0.0
# smtplib2>=0.2.0  # Using built-in smtplib instead
pysnmp>=4.4.12
certbot>=2.7.0
boto3>=1.28.0
cloudflare>=2.11.0
click>=8.1.0
pydantic>=2.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.1.0
# sqlite3 is built into Python
schedule>=1.2.0
jinja2>=3.1.0
marshmallow>=3.20.0
bcrypt>=4.0.0
authlib>=1.2.0
